% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_kill_op}
\alias{add_kill_op}
\title{Replace hvkl operations with harv + kill}
\usage{
add_kill_op(project_path)
}
\arguments{
\item{project_path}{Character string, path to the SWAT+ project folder
(i.e. TxtInOut).}
}
\value{
updated management.sch file
}
\description{
Replace hvkl operations with harv + kill
}
\examples{
\dontrun{
add_kill_op('path_to_your_model')
}
}
