% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_helper.R
\name{add_slope}
\alias{add_slope}
\title{Add condition lines based on slope}
\usage{
add_slope(cond, cond_tbl)
}
\arguments{
\item{cond}{Character string that defines condition for the variable 'var'}

\item{cond_tbl}{Tibble with conditions defined based on soil and land use}
}
\description{
Add condition lines based on slope
}
\keyword{internal}
