% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_suffix_to_duplicate}
\alias{add_suffix_to_duplicate}
\title{Add a running ID to duplicated names}
\usage{
add_suffix_to_duplicate(col_name)
}
\arguments{
\item{col_name}{Character vector of column names}
}
\value{
the \code{col_name} character vector with IDs for duplicated names
}
\description{
Add a running ID to duplicated names
}
\keyword{internal}
