% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_timestamp}
\alias{add_timestamp}
\title{Add a time stamp to the save name}
\usage{
add_timestamp(save_name)
}
\arguments{
\item{save_name}{Name of the saved simulation run}
}
\value{
The name of the simulation run with an added time stamp
}
\description{
Add a time stamp to the save name
}
