% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{aggregate_aa}
\alias{aggregate_aa}
\title{Aggregate and Average Annually}
\usage{
aggregate_aa(tbl)
}
\arguments{
\item{tbl}{A data frame that contains a 'date' column and other numerical columns.}
}
\value{
A numeric vector with the mean of the annual sums for each column in the input data frame.
}
\description{
This function takes a data frame containing a date column and other numerical
columns, aggregates the data annually, and then computes the average of the
annual sums.
}
\examples{
\dontrun{
flo <- aggregate_aa(sim$simulation$flo)
}
}
\keyword{internal}
