% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate time series table}
\usage{
aggregate_time(tbl, time = year, fun = sum, average = FALSE)
}
\arguments{
\item{tbl}{A data frame that contains a 'date' column and multiple numerical
columns.}

\item{time}{A function or a vector of functions to define the time intervals
to which \code{tbl} is aggregated. Must be one or more of the \code{lubridate}
functions \code{year}, \code{month}, \code{day}, or \code{yday}.}

\item{fun}{Aggregation function (e.g. \code{sum}, \code{mean}, \code{min}, or \code{max}).}

\item{average}{Should values be averaged after the aggregation? When e.g.
the time interval is defined by \code{time = c(year, month)}, \code{average = TRUE},
will first aggregate to year and month and then average to mean monthy
values.}
}
\value{
A tibble with the aggregated value columns and added time interval
columns.
}
\description{
\code{aggregate_time()} aggregates tables with a \code{date} column and multiple value
columns to coarser time steps, which are defined by \code{time} using the function
defined by \code{fun}.
}
\examples{
\dontrun{
library(tidyverse)
library(lubridate)

# Generate dummy table

date <- seq(ymd(19700101), ymd(20101231), by = 'day')

n <- 100

tbl <- map(1:n, ~ runif(length(date))) \%>\%
set_names(id_to_run(1:n)) \%>\%
bind_cols(.)

tbl <- bind_cols(date = date, tbl)

# Calculate maximum values for days of the year
aggregate_time(tbl, time = yday, fun = max)

# Calculate mean monthly sums
aggregate_time(tbl, time = c(year, month), fun = sum, average = TRUE)

# Calculate average annual sums
aggregate_time(tbl, time = year, fun = sum, average = TRUE)

}

}
