% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{build_unit_string}
\alias{build_unit_string}
\title{Build the text string with the 'unit' definition to be added to a parameter
name}
\usage{
build_unit_string(vals)
}
\arguments{
\item{vals}{Vector of unit IDs}
}
\value{
A text string with the unit definition.
}
\description{
Build the text string with the 'unit' definition to be added to a parameter
name
}
\keyword{internal}
