% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_fdc}
\alias{calc_fdc}
\title{Calculate Flow Duration Curve (FDC)}
\usage{
calc_fdc(x)
}
\arguments{
\item{x}{a vector or a tibble with flow values.}
}
\value{
a tibble with sorted values and their corresponding exceedance
probabilities.
}
\description{
This function calculates the flow duration curve for a given vector or dataframe.
}
\examples{
\dontrun{
fdc <- calc_fdc(c(3, 1, 4, 1, 5, 9, 2, 6, 5))
}
}
\keyword{internal}
