% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{calc_gof_i}
\alias{calc_gof_i}
\title{Calculate goodness-of-fit values for one function}
\usage{
calc_gof_i(sim, obs, fun)
}
\arguments{
\item{sim}{Data frame with one or many columns with values of the simulated
variable.}

\item{obs}{Data frame with one value column.}

\item{fun}{Goodness-of-fit function}
}
\value{
A vector of goodness-of-fit values for each column of \code{sim}
calculated with \code{fun}
}
\description{
Calculate goodness-of-fit values for one function
}
\keyword{internal}
