% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_segment_diff}
\alias{calc_segment_diff}
\title{Calculate Segment Differences}
\usage{
calc_segment_diff(par, obj, obj_thrs)
}
\arguments{
\item{par}{A list or data frame of parameters to segment.}

\item{obj}{A numeric vector of objective values.}

\item{obj_thrs}{A numeric threshold for the objective values.}
}
\value{
A data frame containing the calculated segment differences.
}
\description{
This function calculates segment differences for a given set of parameters.
}
\keyword{internal}
