% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calc_wyr}
\alias{calc_wyr}
\title{Calculate water yield ratio}
\usage{
calc_wyr(sim)
}
\arguments{
\item{sim}{A SWATrunR simulation run with the above listed simulated
variables.}
}
\value{
The water yield ratio.
}
\description{
Calculate the water yield ratio from simulated average annual water balance
components. The simulation run must provide
\describe{
\item{precip}{Variable out of 'basin_wb_aa' output file.}
\item{surq_cha}{Variable out of 'basin_wb_aa' output file.}
\item{surq_res}{Variable out of 'basin_wb_aa' output file.}
\item{latq_cha}{Variable out of 'basin_wb_aa' output file.}
\item{latq_res}{Variable out of 'basin_wb_aa' output file.}
\item{qtile}{Variable out of 'basin_wb_aa' output file.}
\item{flo}{Variable out of ''basin_aqu_aa' output file.}
}
}
