% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{calculate_performance}
\alias{calculate_performance}
\title{Calculate All Performance Metrics}
\usage{
calculate_performance(
  sim,
  obs,
  par_name = NULL,
  perf_metrics = NULL,
  period = NULL,
  fn_summarize = "mean"
)
}
\arguments{
\item{sim}{Object from SWATrunR}

\item{obs}{Dataframe for the observed data with two columns: date and value.}

\item{par_name}{(optional) Character for name of the parameter set to be used.
Default \code{par_name = NULL}, which select first variable in sim object.
But if you have multiple parameter sets, you can use this argument to select
the one you want to use. Example par_name = "flo_day" or par_name = no3_day_conc".}

\item{perf_metrics}{(optional) Character vector with the names of the performance metrics
to used in the rank_tot calculation. Default \code{perf_metrics = NULL},
which means that all performance metrics will be used in calculation.
Other example could be perf_metrics = c("kge", "nse"), which means that only
KGE and NSE will be used in calculation.}

\item{period}{(optional) Character describing, which time interval to display.
Default \code{period = NULL}, which mean not activated, other examples are
"day", "week", "month", etc).
See \href{https://www.rdocumentation.org/packages/lubridate/versions/1.3.3/topics/floor_date}{lubridate::floor_date} for details.}

\item{fn_summarize}{(optional) Function to recalculate to the specified time interval.
Default \code{fn_summarize ="mean"}, other examples are "median", "sum". See \href{https://dplyr.tidyverse.org/reference/summarise.html}{dplyr::summarise} for details.}
}
\value{
A dataframe with the performance metrics and ranking.
}
\description{
This function calculates various performance metrics for a given simulation
and observed data.
}
\examples{
\dontrun{
obj_tbl <- calculate_performance(sim = sim_flow, obs, "flo_day", c("kge", "nse"), "month", "sum")
}
}
\seealso{
\code{\link{calculate_performance_2plus}}
}
\keyword{calculate}
