% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{calculate_wyr}
\alias{calculate_wyr}
\title{Calculate Water Yield Ratio (WYR)}
\usage{
calculate_wyr(sim)
}
\arguments{
\item{sim}{A SWATrunR nested list containing simulation data. It should have a named element
'simulation' which itself contains named elements: 'precip', 'surq_cha', 'surq_res',
'latq_cha', 'latq_res', 'qtile', and 'flo'.}
}
\value{
A numeric value representing the Water Yield Ratio (WYR).
}
\description{
This function calculates the Water Yield Ratio (WYR) based on various components
of a simulation, including precipitation, surface runoff, lateral flow, and
others. It aggregates these components and computes the ratio of the sum of
water yields to the precipitation.
}
\examples{
\dontrun{
# Calculate WYR
wyr <- calculate_wyr(sim)
}
}
\keyword{calculate}
