% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_date_col}
\alias{check_date_col}
\title{Check and fix position of date column in a table}
\usage{
check_date_col(tbl)
}
\arguments{
\item{tbl}{Data frame which for which date column is checked and fixed.}
}
\value{
The \code{tbl} where the date column is put into first position and
renamed to 'date'
}
\description{
Check and fix position of date column in a table
}
\keyword{internal}
