% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_point_shape}
\alias{check_point_shape}
\title{Check a point value shape value if it is a correct value between 0 and 25}
\usage{
check_point_shape(shape)
}
\arguments{
\item{shape}{Point shape value. To be a correct value it must be one of 0:25.
Any other value returns NA}
}
\value{
The correct point shape value or NA.
}
\description{
Check a point value shape value if it is a correct value between 0 and 25
}
\keyword{internal}
