% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{copy_file_version}
\alias{copy_file_version}
\title{Copy a file with an optional version suffix.}
\usage{
copy_file_version(model_path, file_name, file_version = NULL)
}
\arguments{
\item{model_path}{The path to the directory containing the file.}

\item{file_name}{The name of the file to be copied.}

\item{file_version}{An optional version suffix to be appended to the filename.
Default \code{file_version = NULL}, which adds '.bkp' to the filename.}
}
\value{
None
}
\description{
This function copies a file from one location to another, with an optional
version suffix appended to the filename.
}
