% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cuts_to_rect}
\alias{cuts_to_rect}
\title{Convert Cuts to Rectangle Coordinates}
\usage{
cuts_to_rect(lbl)
}
\arguments{
\item{lbl}{A character vector of cut labels.}
}
\value{
A data frame containing the rectangle coordinates.
}
\description{
This function converts cut labels into rectangle coordinates.
}
\keyword{internal}
