% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dotty_fig}
\alias{dotty_fig}
\title{Make a dotty plot}
\usage{
dotty_fig(sim_yield, yld_obs = NULL)
}
\arguments{
\item{sim_yield}{a dataframe with yield data}

\item{yield}{(optional) The mean yield values for each crop.
Default \code{yield = NULL}.}

\item{yield_min}{(optional) The minimum observed yield values for each crop.
Default \code{yield_min = NULL}.}

\item{yield_max}{(optional) The maximum observed yield values for each crop.
Default \code{yield_max = NULL}.}
}
\value{
ggplot object for dotty plot
}
\description{
Make a dotty plot
}
\examples{
\dontrun{
dotty_fig(sim_yield)
}
}
\keyword{internal}
\keyword{plot}
