% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{ensure_dir}
\alias{ensure_dir}
\title{Ensure Directory Exists}
\usage{
ensure_dir(path)
}
\arguments{
\item{path}{A character string specifying the path to the directory.}
}
\value{
Invisibly returns the directory path as a character string.
}
\description{
Checks if a directory exists at the specified path and creates it if it does not exist.
This is useful to make sure folders exist before writing files or creating subdirectories.
}
\keyword{internal}
