% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_par_range}
\alias{find_par_range}
\title{Find Parameter Range for Simulation}
\usage{
find_par_range(par, sim, obj, rel_rng)
}
\arguments{
\item{par}{A numeric vector of parameter values.}

\item{sim}{A numeric vector of simulation results corresponding to the parameter values.}

\item{obj}{A numeric value representing the objective value to compare against.}

\item{rel_rng}{A numeric value specifying the relative range for acceptable error.}
}
\value{
A list with the following components:
\describe{
\item{par_rng}{A numeric vector of length 2 containing the minimum and maximum parameter values
within the specified relative range of the objective.}
\item{par_val}{A numeric value of the parameter that minimizes the absolute error relative to the objective.}
\item{x}{A numeric vector of interpolated parameter values.}
\item{y}{A numeric vector of interpolated simulation results corresponding to \code{x}.}
}
}
\description{
This function identifies the range of parameter values that produce simulation results
within a specified relative range of a target objective value. It also returns the parameter
value that minimizes the absolute error relative to the objective.
}
\examples{
\dontrun{
find_par_range(sim[[1]], calculate_wyr(sim), 0.27, 0.05)
}
}
\keyword{internal}
