% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fix_dates}
\alias{fix_dates}
\title{Fix the dates}
\usage{
fix_dates(runr_obj, obs_obj, trim_start = NULL, trim_end = NULL)
}
\arguments{
\item{runr_obj}{SWATrunR object}

\item{obs_obj}{Observation dataframe with two columns named "date" and "value"}

\item{trim_start}{(optional) The starting date to trim the data times series.
Default \code{trim_start = NULL}. Example: trim_start = "2010-01-01"}

\item{trim_end}{(optional) The ending date to trim the data times series.
Default \code{trim_end = NULL}. Example trim_end = "2015-01-01"}
}
\value{
list with two elements: "simulation" SWATrunR object and
"observation" dataframe with the same dates period.
}
\description{
Adjust the periods of the observation and simulation data so that they match.
}
\examples{
\dontrun{
my_obj <- fix_dates(runr_obj = sim_flow, obs_obj = obs_wq,
trim_start = "2010-01-01", trim_end = "2015-01-01")
}
}
\keyword{data}
\keyword{manipulation}
