% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{get_conc}
\alias{get_conc}
\title{Calculate concentrations}
\usage{
get_conc(sim, not_conc = "^flo_day|^gwd", sediment = "^sed_")
}
\arguments{
\item{sim}{A simulation object containing the data.}

\item{not_conc}{(optional) Regular expression pattern for variables that should not be
considered for concentration calculation. Default \code{not_conc = '^flo_day|^gwd'}.}

\item{sediment}{(optional) Regular expression pattern for sediment variables.
Default \code{sediment = '^sed_'}.}
}
\value{
A modified simulation object with added concentration variables.
}
\description{
This function calculates concentrations based on simulated data, adjusting
for specified parameters.
}
\examples{
\dontrun{
# Example usage:
sim_data <- get_conc(sim_data)
}
}
\keyword{data}
\keyword{manipulation}
