% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{group_hydr_values}
\alias{group_hydr_values}
\title{Group the values of a parameter in the file hydrology.hyd.}
\usage{
group_hydr_values(par_name, model_path)
}
\arguments{
\item{par_name}{Name of the parameter for which values should be grouped}

\item{model_path}{path to the SWAT+ model project folder which contains the
file hydrology.hyd}
}
\value{
A vector with unique indices of the parameter value groups assigned
to each HRU.
}
\description{
This is useful for applying different parameter ranges e.g. to parameters
such as perco, cn3_swf, or latq_co which have different initial values
based on runoff or leaching potential.
}
