% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{group_values}
\alias{group_values}
\title{ID Grouping Functions for Parameter Names}
\usage{
group_values(vals, sep = ":", n_trunc = Inf)
}
\arguments{
\item{vals}{A vector of values to be grouped. Can be numeric or character.}

\item{sep}{(optional) A string separator to use between ranges of numeric
values. Default \code{sep = ':'}.}

\item{n_trunc}{(optional) An integer specifying the maximum number of elements
to include in the output. Default \code{sep = Inf}, which means no truncation.}
}
\value{
A string with grouped values. For numeric inputs, ranges are collapsed
and separated by the specified separator. Non-numeric inputs are concatenated
with ', '. If all values are NA, an empty string is returned.
}
\description{
This function groups values for parameter names, sorting and concatenating
numeric sequences while handling non-numeric values appropriately.
}
\keyword{internal}
