% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{guess_time_step}
\alias{guess_time_step}
\title{Guess the time step of a vector of values (e.g. years, days,...)}
\usage{
guess_time_step(x)
}
\arguments{
\item{x}{Numeric vector with year, month, or day values}
}
\value{
A text string indicating the guessed time interval
}
\description{
Guess the time step of a vector of values (e.g. years, days,...)
}
\keyword{internal}
