% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{id_to_run}
\alias{id_to_run}
\title{Convert run ID values to run names}
\usage{
id_to_run(run_ids, id_max = NULL)
}
\arguments{
\item{run_ids}{Integer vector with the run IDs}

\item{id_max}{Largest value for run IDs (defines the number of leading zeros)}
}
\value{
An character vector with the run names
}
\description{
Convert run ID values to run names
}
\examples{
\dontrun{
run_sel <- c(1,36, 598, 2311)
id_to_run(run_sel)
}

}
