% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{id_to_unit}
\alias{id_to_unit}
\title{Generate 'unit' condition for parameter definition from ID vector}
\usage{
id_to_unit(ids)
}
\arguments{
\item{ids}{Numeric ID vector.}
}
\value{
Character string with unit condition based on the given IDs.
}
\description{
\code{id_to_unit()} generates a 'unit' condition for  \code{SWATrunR} parameter
definitions from an ID vector. This is useful when e.g. a vector of HRU IDs
is given and a parameter change change should be defined for those units
}
\examples{
\dontrun{
# HRU IDs e.g. same land use
hru_ids <- c(1, 3, 4, 5, 28, 37:43, 58)

# E.g. define a change of cn2
cn2_chg <- c('cn2_agri::cn2.hru | change = abschg' = - 5)

# Add unit condition
names(cn2_chg) <- paste(names(cn2_chg), id_to_unit(hru_ids))
}
}
