% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_helper.R
\name{init_cal}
\alias{init_cal}
\title{Initialize the calibration.cal table}
\usage{
init_cal(par_def)
}
\arguments{
\item{par_def}{Tibble with one line that includes the parameter definition of
parameter i}
}
\description{
Initialize the calibration.cal table
}
\keyword{internal}
