% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_workflow - Copy.R
\name{initialize_cropcal}
\alias{initialize_cropcal}
\title{Initialize a hard crop calibration workflow}
\usage{
initialize_cropcal(project_name, path, model_path, start_new_session = FALSE)
}
\arguments{
\item{project_name}{Name of the crop calibration project.}

\item{path}{Path where the crop calibration project should be initialized.}

\item{model_path}{Path to the SWAT+ model.}

\item{start_new_session}{Open R project in new or current R session.}
}
\value{
Opens a template R project for SWAT+ crop calibration.
}
\description{
\code{initialize_cropcal()} initializes a template crop calibration project. The
project provides a template workflow in the folder
'/workflow'. The folders '/observation' and '/simulation' are generated
for further use in the simulation workflow.
An R project is initialized and loaded in the new folder structure.
}
\examples{
\dontrun{
initialize_cropcal('cropcal_test', '~/', 'Path:/to/txt_inout')
}

}
