% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_workflow.R
\name{initialize_hardcal}
\alias{initialize_hardcal}
\title{Initialize a hard calibration workflow}
\usage{
initialize_hardcal(project_name, path, model_path, start_new_session = FALSE)
}
\arguments{
\item{project_name}{Name of the hard calibration project.}

\item{path}{Path where the hard calibration project should be initialized.}

\item{model_path}{Path to the SWAT+ model.}

\item{start_new_session}{Open R project in new or current R session.}
}
\value{
Opens a template R project for SWAT+ hard calibration.
}
\description{
\code{initialize_hardcal()} initializes a template hard calibration project. The
template project provides a template simulation workflow in the folder
'/workflow'. The folders '/observation' and '/simulation' are generated
for further use in the simulation workflow.
An R project is initialized and loaded in the new folder structure.
}
\examples{
\dontrun{
initialize_hardcal('hardcal_test', '~/', 'Path:/to/txt_inout')
}

}
