% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_workflow.R
\name{initialize_softcal}
\alias{initialize_softcal}
\title{Initialize a soft calibration workflow}
\usage{
initialize_softcal(project_name, path, model_path, start_new_session = FALSE)
}
\arguments{
\item{project_name}{Name of the soft calibration project.}

\item{path}{Path where the soft calibration project should be initialized.}

\item{model_path}{Path to the SWAT+ model.}

\item{start_new_session}{Open R project in new or current R session.}
}
\value{
Opens a template R project for SWAT+ soft calibration.
}
\description{
\code{initialize_softcal()} initializes a template soft calibration project. The
project provides a template workflow in the folder '/workflow'. The folders
'/observation' and '/simulation' are generated for further use in the
simulation workflow. An R project is initialized and loaded in the new folder
structure.
}
\examples{
\dontrun{
initialize_softcal('softcal_test', '~/', 'Path:/to/txt_inout')
}

}
