% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{insert_line_at}
\alias{insert_line_at}
\title{Insert line into character with multiple lines}
\usage{
insert_line_at(dat, txt, insert_after)
}
\arguments{
\item{dat}{character with multiple lines}

\item{txt}{character line to be inserted}

\item{insert_after}{numeric for line number after, which to insert txt}
}
\value{
character with multiple lines with inserted new lines
}
\description{
Insert line into character with multiple lines
}
\keyword{internal}
