% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{load_to_conc}
\alias{load_to_conc}
\title{Convert load and discharge time series into a concentration time series.}
\usage{
load_to_conc(
  load,
  flow,
  load_unit = "kg day-1",
  flow_unit = "m3 s-1",
  out_unit = "mg L-1"
)
}
\arguments{
\item{load}{Data frame with a date column and one or many columns with load
values. The number of load columns must be the same as the number of
discharge columns in the data frame \code{flow}.}

\item{flow}{Data frame with a date column and one or many columns with
discharge values. The number of flow columns must be the same as the number
of load columns in the data frame \code{load}.}

\item{load_unit}{Unit of the load values provided as a text string in the
format \verb{'mass time-1}. Default is \code{'kg day-1'}.}

\item{flow_unit}{Unit of the discharge values provided as a text string in
the format \verb{'volume time-1}. Default is \code{'m3 s-1'}.}

\item{out_unit}{Unit of the concentration values provided as a text string in
the format \verb{'mass volume-1}. Default is \code{'mg L-1'}.}
}
\value{
A table with date column and calculated concentration values.
}
\description{
Calculate a time series of concentration values from time series of
load and discharge.
}
\examples{
\dontrun{
# Example usage:
date_load <- seq(as.Date('2000-01-01'), as.Date('2000-12-31'), by = '2 week')
date_flow <- seq(as.Date('2000-01-01'), as.Date('2000-12-31'), by = 'day')

l_df <- data.frame(date = date_load,
                   value = runif(length(date_load)))

f_df <- data.frame(date = date_flow,
                   value = runif(length(date_flow)))

conc <- load_to_conc(load = l_df, flow = f_df, load_unit = 'tons day-1')
}
}
\keyword{data}
\keyword{manipulation}
