% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{normalize}
\alias{normalize}
\title{Normalize values in vector between 0 and 1}
\usage{
normalize(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A vector with normalized values.
}
\description{
Normalize values in vector between 0 and 1
}
\keyword{internal}
