% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_calval_comparison}
\alias{plot_calval_comparison}
\title{Plot Calibration-Validation Comparison}
\usage{
plot_calval_comparison(cal_tbl, val_tbl, indexes = NULL)
}
\arguments{
\item{cal_tbl}{A data frame containing calibration performance results.}

\item{val_tbl}{A data frame containing validation performance results.}

\item{indexes}{(optional) An optional vector of metric names to filter and
include in the comparison plot. Default \code{indexes = NULL} includes all
metrics.}
}
\value{
A ggplot object representing the comparison between calibration and
validation results for the specified metrics.
}
\description{
This function takes calibration and validation tables and produces a
comparison boxplot for various metrics.
}
\examples{
\dontrun{
plot_calval_comparison(obj_tbl_cal, obj_tbl_val, indexes = c("nse", "pbias"))
}
}
\seealso{
\code{\link{calculate_performance}}, \code{\link{calculate_performance_2plus}}
}
\keyword{plot}
