% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dotty}
\alias{plot_dotty}
\title{Plot Dotty}
\usage{
plot_dotty(
  par,
  var,
  y_label = "y",
  n_col = 3,
  y_lim = NULL,
  y_inter = NULL,
  trend = FALSE,
  run_ids = NULL,
  low_up = FALSE
)
}
\arguments{
\item{par}{A data frame of model parameter values for each simulation.}

\item{var}{The model performance result vector or list of vectors to be
plotted against parameter values.}

\item{y_label}{(optional) Labels for the y-axis, either a single label or a vector
corresponding to each variable if \code{var} is a list. Default \code{y_label = 'y'}.}

\item{n_col}{(optional) Number of columns in the facet grid. Default \code{n_col = 3}.}

\item{y_lim}{(optional) Limits for the y-axis. Default \code{y_lim = NULL}.}

\item{y_inter}{(optional) Y-axis intercept value. Default \code{y_inter = NULL}.}

\item{trend}{(optional) Logical, indicating whether to add a trend line.
Default is \code{trend = FALSE}.}

\item{run_ids}{(optional) A numeric vector of run IDs to be highlighted in the plot.
Default is \code{run_ids = NULL}.}

\item{low_up}{(optional) Logical, TRUE if whole possible parameter range should be used
for x axis. Default is \code{low_up = FALSE}.}
}
\value{
A ggplot object representing the dotty plot.
}
\description{
This function creates a dotty plot to visualize the relationship between
parameter values and performance results.
}
\examples{
# Example usage:
# plot_dotty(my_data, 'performance', y_label = 'y_label')

}
\keyword{plot}
