% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dotty_yields}
\alias{plot_dotty_yields}
\title{Plot Dotty Yields}
\usage{
plot_dotty_yields(sim_result, yield = NULL, yield_min = NULL, yield_max = NULL)
}
\arguments{
\item{sim_result}{The simulation results containing yield data.}

\item{yield}{(optional) The mean yield values for each crop.
Default \code{yield = NULL}.}

\item{yield_min}{(optional) The minimum observed yield values for each crop.
Default \code{yield_min = NULL}}

\item{yield_max}{(optional) The maximum observed yield values for each crop.
Default \code{yield_max = NULL}}
}
\value{
A combined ggplot object showing dottty plots for 4 parameters.
}
\description{
This function creates dotty plots for simulated yields, comparing them with
observed yields (if available).
}
\examples{
\dontrun{
plot_phu_yld_bms(ylds_phu_dmat, dmat_chg, yield_mean, yield_min, yield_max)
}
}
\keyword{plot}
