% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_esco_range}
\alias{plot_esco_range}
\title{Plot ESCO Range}
\usage{
plot_esco_range(sim, obs_wy_ratio, rel_wyr_lim = 0.05)
}
\arguments{
\item{sim}{The simulation results water balance data. Including:
\describe{
\item{precip}{Variable out of 'basin_wb_day' output file.}
\item{surq_cha}{Variable out of 'basin_wb_day' output file.}
\item{surq_res}{Variable out of 'basin_wb_day' output file.}
\item{latq_cha}{Variable out of 'basin_wb_day' output file.}
\item{latq_res}{Variable out of 'basin_wb_day' output file.}
\item{qtile}{Variable out of 'basin_wb_day' output file.}
\item{flo}{Variable out of ''basin_aqu_day' output file.}
}}

\item{obs_wy_ratio}{A numeric value representing the observed water yield ratio.}

\item{rel_wyr_lim}{A numeric value specifying the relative range for acceptable error.}
}
\value{
A ggplot object.
}
\description{
This function creates a ggplot visualization of the ESCO range data,
highlighting specific values and limits with horizontal and vertical lines.
}
\examples{
\dontrun{
plot_esco_range(sim_esco, obs_wy_ratio)
}
}
