% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_oat}
\alias{plot_oat}
\title{Function to plot OAT (One-At-A-Time) analysis results}
\usage{
plot_oat(sim, obs = NULL, variable, round_values = 3)
}
\arguments{
\item{sim}{Object containing simulation data from SWATrunR.}

\item{obs}{Optional dataframe for observed data with columns 'date' and 'value'. Default is NULL.}

\item{variable}{Name of the variable for which OAT analysis results are plotted.}

\item{round_values}{Number of decimal places to round parameter values. Default is 3.}
}
\value{
A dygraph object showing the OAT analysis results.
}
\description{
This function generates an interactive dygraph to visualize the results of an OAT analysis,
comparing simulated data with observed data if provided.
}
\examples{
\dontrun{
plot_oat(sim_oat, obs = obs_data, variable = 'flo_day', round_values = 2)
}
}
\seealso{
\code{\link{plot_timeseries}}, \code{\link{sample_oat}}, \code{\link[SWATrunR:run_swatplus]{https://chrisschuerz.github.io/SWATrunR/}}
}
