% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_phu_yld_bms}
\alias{plot_phu_yld_bms}
\title{Plot PHU, yield, and biomass over days to maturity}
\usage{
plot_phu_yld_bms(sim_result, yield_obs = NULL, bar_width = 0.5)
}
\arguments{
\item{sim_result}{The simulation results containing PHU, yield, and biomass
data.}

\item{yield_obs}{(optional) A data frame with observed minimum, maximum, and mean yield
values for each crop. Default is \code{NULL}. No data is provided. If provided,
the data frame should contain the following columns: \code{plant_name},
\code{yield_min}, \code{yield_max}, and \code{yield_mean}.}

\item{bar_width}{(optional) Value from 0 to 1 controls the width of the box
in the boxplot. Default \code{bar_width = 0.5}.}
}
\value{
A combined ggplot object showing boxplots for PHU fractions, yields,
and biomass.
}
\description{
This function generates boxplots to visualize Plant Heat Units (PHU)
fractions, yields, and biomass over changes in days to maturity.
}
\keyword{plot}
