% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_selected_sim}
\alias{plot_selected_sim}
\title{Plot interactive time series of simulation results}
\usage{
plot_selected_sim(
  sim,
  obs = NULL,
  par_name = NULL,
  run_ids = NULL,
  run_sel = NULL,
  plot_bands = TRUE,
  period = NULL,
  fn_summarize = "mean",
  x_label = "Date",
  y_label = "Discharge (m<sup>3</sup> s<sup>-1</sup>)"
)
}
\arguments{
\item{sim}{Object from SWATrunR.}

\item{obs}{(optional) Dataframe for observed data with columns 'date' and
'value'. Default \code{obs = NULL}, only simulation data are available.}

\item{par_name}{(optional) Name of the parameter set to use.
Default \code{par_name = NULL} uses the first variable in sim object.
For multiple parameter sets, specify the one to use, e.g.,
par_name = "flo_day" or par_name = "no3_day_conc".}

\item{run_ids}{(optional) Character vector of run IDs to plot.
Default \code{run_ids = NULL} requires input of run_sel.}

\item{run_sel}{(optional) Character vector of run IDs to emphasize in the plot.
Default \code{run_sel = NULL} requires input of run_ids.}

\item{plot_bands}{(optional) Logical. If TRUE, includes lower and upper bands of results.
Default \code{plot_bands = TRUE}.}

\item{period}{(optional) Character defining the time step for aggregation. Options include "day", "week",
"month", "year", etc. Use "average monthly" for multi-annual monthly values.
Default \code{period = NULL} provides no aggregation. See \href{https://www.rdocumentation.org/packages/lubridate/versions/1.3.3/topics/floor_date}{lubridate::floor_date} for details.}

\item{fn_summarize}{Function to summarize time intervals. Default \code{fn_summarize ="mean"}, other examples are "median", "sum".
See \href{https://dplyr.tidyverse.org/reference/summarise.html}{dplyr::summarise} for details.}

\item{x_label}{(optional) Character for x axis label. Default \code{x_label = "Date"}.}

\item{y_label}{(optional) Character for y axis label. Default \code{y_label = "Discharge (m³/s)"}.}
}
\value{
Interactive plot of simulated and observed data.
}
\description{
This function plots interactive time series comparing simulated data with observed data.
}
\examples{
\dontrun{
plot_selected_sim(sim_flow, obs = obs, run_sel = c(95), run_ids = run_sel_ids,
period = "average monthly", plot_bands = TRUE)
}
}
\keyword{plot}
