% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time series of simulation results and observations}
\usage{
plot_timeseries(
  sim,
  obs = NULL,
  run_sel = NULL,
  run_ids = NULL,
  plot_bands = TRUE,
  sim_pointshape = 1,
  obs_pointshape = 1,
  sim_linetype = "solid",
  obs_linetype = "dotted",
  run_sel_color = "#A50F15",
  run_ids_color = NULL,
  obs_color = "black",
  band_color = "#CB181D",
  band_alpha = 0.2,
  run_sel_label = "Simulation",
  obs_label = "Observation",
  band_label = "Simulated range",
  x_label = "Date",
  y_label = expression(Discharge ~ (m^3 ~ s^{
     -1
 })),
  legend_pos = c(1, 0.99, 0.99),
  split_years = NULL,
  free_y = FALSE
)
}
\arguments{
\item{sim}{Data frame with one date column and one or many columns with
values of the simulated variable.}

\item{obs}{(optional) Data frame with one date and one value column.
Default \code{obs = NULL}, only simulations are plotted.}

\item{run_sel}{(optional) Integer value or name of simulation run which
should be plotted as 'selected' run. Default \code{run_sel = NULL} no run
is plotted as 'selected' run. In this case \code{run_ids} must be selected for
plotting.}

\item{run_ids}{(optional) Integer vector or character vector of run IDs
or run names to plot. Default \code{run_ids = NULL}. In this case only the
'seleced' simulation run from \code{run_sel} is plotted. If \code{plot_bands = TRUE}
the selected \code{run_ids} are used to calculate and plot the upper and lower
boundaries of the plotted band. If \code{plot_bands = FALSE} a maximum number of
10 \code{run_ids} can be selected for plotting.}

\item{plot_bands}{(optional) If TRUE, upper and lower boundary values are
calculated from the selected \code{run_ids} and a band is plotted for them. If
FALSE the individual selected runs from \code{run_ids} are plotted.
Default \code{plot_bands = TRUE}.}

\item{sim_pointshape}{(optional) Integer value defining the shape of the points to be
plotted for the simulated time series of \code{run_sel}. The value must be a
value between 0 and 25. For all other inputs no points will be plotted.
Default is \code{sim_pointshape = 1}.}

\item{obs_pointshape}{(optional) Integer value defining the shape of the points to be
plotted for the observation time series. The value must be a
value between 0 and 25. For all other inputs no points will be plotted.
Default is \code{obs_pointshape = 1}.}

\item{sim_linetype}{(optional) Type of the line to be plotted for the
simulated time series. The linetype must be one of 'blank', 'solid',
'dashed', 'dotted', 'dotdash', 'longdash\verb{, or 'twodash'. When 'blank' no line will be plotted. Default is }sim_linetype = 'solid'`.}

\item{obs_linetype}{Type of the line to be plotted for the
observation time series. The linetype must be one of 'blank', 'solid',
'dashed', 'dotted', 'dotdash', 'longdash\verb{, or 'twodash'. When 'blank' no line will be plotted. Default is }obs_linetype = 'dotted'`.}

\item{run_sel_color}{(optional) Color of the line and points plotted for the
simulation time series \code{run_sel}. Default \code{run_sel_color = '#A50F15'}.}

\item{run_ids_color}{(optional) Colors of the lines plotted for the
simulation time series selected with \code{run_ids}. The colors are only used
when \code{plot_bands = FALSE} and individual timeseries are plotted. The color
vector must be of the same length as the selected runs in \code{run_ids}.
Default \code{run_ids_color = NULL} where a default color pallete is used.}

\item{obs_color}{(optional) Color of the line and points of the plotted
observation time series. Default \code{obs_color = 'black'}.}

\item{band_color}{(optional) Color of the band plotted fur the \code{run_ids}. The color is
only used when \code{plot_bands = TRUE}. Default \code{band_color = '#CB181D'}.}

\item{band_alpha}{(optional) Transparency value of the band plotted fur the
\code{run_ids}. The value is only used when \code{plot_bands = TRUE}.
The value must be set between 0 and 1. Default \code{band_alpha = 0.2}.}

\item{run_sel_label}{(optional) Label which is plotted in the legend to
indicate the selected simulation run \code{run_sel}. Default is
\code{run_sel_label = 'Simulation'}.}

\item{obs_label}{(optional) Label which is plotted in the legend to
indicate the observation time series. Default is
\code{run_obs = 'Observation'}.}

\item{band_label}{(optional) Label which is plotted in the legend to
indicate the simulation band for the \code{run_ids}. Default is
\code{band_label = 'Simulated range'}.}

\item{x_label}{(optional) x-Axis label. Default is \code{x_label = 'Date'}.}

\item{y_label}{(optional) y-Axis label. Default is \code{x_label = expression(Discharge~(m^3~s^{-1}))}.}

\item{legend_pos}{(optional) Position of the legend inside of the plot panel.}

\item{split_years}{(optional) Integer value which is used to split the
time series into time intervals of the length \code{split_years} to plot those
time periods in separate plot panels. This has an advantage for plotting
long time series. Default \code{split_years = NULL} and the entire time series
is plotted in one plot panel.}

\item{free_y}{If \code{split_years} is defined the y-Axis can be scaled individually
for all plot panels if \code{free_y = TRUE}. Default \code{free_y = FALSE} and all
plot panels have the same y plot range.}
}
\value{
A ggplot object of the plotted time series.
}
\description{
\code{plot_timeseries()} plots a time series comparing simulated data with observed data.
}
\keyword{plot}
