% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare_plant_parameter}
\alias{prepare_plant_parameter}
\title{Rearrange parameter change tables to SWATrunR format}
\usage{
prepare_plant_parameter(par_tbl)
}
\arguments{
\item{par_tbl}{A table with parameter changes where the first column is
'plant_name' and the following columns are the parameters change definitions.}
}
\value{
The resturctured SWATrunR parameter table
}
\description{
Rearrange parameter change tables to SWATrunR format
}
