% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{rank_gof}
\alias{rank_gof}
\title{Rank lines in a table of goodness-of-fit values.}
\usage{
rank_gof(gof_tbl, weights = NULL, type = "rank")
}
\arguments{
\item{gof_tbl}{Table with goodness-of-fit values, where each column provides
the values for all \code{run}s of one GOF index}

\item{weights}{(optional) weights vector to apply different weights to
the goodness-of-fit indices for the calculation of the total rank.}

\item{type}{Either \code{'rank'} to calculate absolute ranks or \code{'norm'} to use
normalized values of the indices.}
}
\value{
A rank table for all goodness-of-fit columns and a total rank.
}
\description{
\code{rank_gof()} ranks each column of a goodness-of-fit table either by absolute
ranks or normalized GOF values and calculates a total rank for each line.
}
