% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_helper.R
\name{read_unit_conditions}
\alias{read_unit_conditions}
\title{Read the unit numbers (for hru, aqu, cha, res) and the textures etc for later
parameter conditioning.}
\usage{
read_unit_conditions(project_path, parameter)
}
\arguments{
\item{project_path}{Path to the SWAT+ project}
}
\description{
Read the unit numbers (for hru, aqu, cha, res) and the textures etc for later
parameter conditioning.
}
\keyword{internal}
