% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{remove_unsuccesful_runs}
\alias{remove_unsuccesful_runs}
\title{Remove Unsuccessful Runs from Simulation}
\usage{
remove_unsuccesful_runs(sim)
}
\arguments{
\item{sim}{A SWATrunR nested list containing simulation data. : \code{simulation} and \code{parameter}.
\itemize{
\item \code{simulation}: A list where the first element contains the results of the simulation runs.
\item \code{parameter}: A list with an element \code{values} which is a data frame containing the parameter sets used in the simulation.
}}
}
\value{
A modified version of the input \code{sim} list, with the \code{parameter$values}
element containing only the parameter sets for which simulation results are available.
}
\description{
This function removes unsuccessful runs from the simulation parameter set
to ensure that all analyses are only performed for parameter combinations
that have corresponding simulation results.
}
\examples{
\dontrun{
sim_flow_full <- remove_unsuccesful_runs(sim_flow_full_bck)
}
}
\keyword{helper}
