% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_closest_value}
\alias{replace_closest_value}
\title{Replace Closest Value in Sequence}
\usage{
replace_closest_value(val_seq, x)
}
\arguments{
\item{val_seq}{A numeric vector representing a sequence of values.}

\item{x}{A numeric value to find the closest match in the sequence.}
}
\value{
A numeric vector with the closest value replaced by \code{x}.
}
\description{
Replaces the closest value in a sequence with a specified value.
}
\keyword{internal}
