% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{replace_col_by_sequence}
\alias{replace_col_by_sequence}
\title{Replace Column by Sequence}
\usage{
replace_col_by_sequence(tbl, val_seq, i)
}
\arguments{
\item{tbl}{A data frame or matrix.}

\item{val_seq}{A numeric vector representing a sequence of values.}

\item{i}{An integer indicating the index of the column to be replaced.}
}
\value{
The modified data frame or matrix with the specified column replaced
by \code{val_seq}.
}
\description{
Replaces the specified column in a table with a sequence of values.
}
\keyword{internal}
