% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{sample_days_mat}
\alias{sample_days_mat}
\title{Sample days to maturity values}
\usage{
sample_days_mat(
  crop_names,
  change_min = -60,
  change_max = 90,
  change_step = 10
)
}
\arguments{
\item{crop_names}{Character vector with names of selected crops. All crops
must be defined in the plants.plt backup file saved in ./backup/plants.plt}

\item{change_min}{Minimum value of change for the days to maturity. Default
is -60.}

\item{change_max}{Minimum value of change for the days to maturity. Default
is 90.}

\item{change_step}{Minimum value of change for the days to maturity. Default
is 10.}
}
\value{
A parameter input table for the input argument \code{parameter} of the
\code{SWATrunR} function \code{run_swatplus()}.
}
\description{
\code{sample_days_mat()} reads the initial days to maturity values for the
selected \code{crop_names} from the unchanged plants.plt file (saved in
./backup/plants.plt) and generates a parameter table which can be used with
the \code{SWATrunR} function \code{run_swatplus()} perform simulations with changed
days to maturity values for the selected crops.
}
\examples{
\dontrun{
library(SWATtunR)

# Crops for which days to maturity values are sampled
crop_names <- c('csil', 'wbar', 'sgbt', 'wwht')

# Generate a days to maturity parameter table
par_dmat <- sample_dmat(crop_names)

par_dmat
}
}
