% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{sample_lhs}
\alias{sample_lhs}
\title{Sample parameters space}
\usage{
sample_lhs(par, n)
}
\arguments{
\item{par}{A matrix or data frame specifying the parameter ranges.}

\item{n}{The number of samples to generate.}
}
\value{
A tibble containing the sampled parameter values.
}
\description{
This function samples the parameter space using Latin Hypercube Sampling (LHS).
}
\examples{
# Define parameter ranges
parameters <- data.frame(
  param1 = c(0, 1),
  param2 = c(0, 2),
  param3 = c(0, 10)
)

# Sample the parameter space
sample_space <- sample_lhs(parameters, 3)
sample_space
}
\keyword{sampling}
