% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{sample_oat}
\alias{sample_oat}
\title{Sample parameter values one-at-a-time (OAT)}
\usage{
sample_oat(par, par_center = 1, n_t = 10)
}
\arguments{
\item{par}{A matrix or data frame specifying the parameter ranges.
Each element is length 2 indicating the min and max values.}

\item{par_center}{(optional) A data frame or matrix of parameter centers.
Default \code{par_center = 1}.}

\item{n_t}{(optional) An integer specifying the parameter combination around,
which you want to do the OAT analysis. Default \code{n_t = 10}.}
}
\value{
A data frame containing the OAT samples.
}
\description{
Generates a set of OAT (One-At-a-Time) samples based on provided parameter centers and boundaries.
}
\examples{
\dontrun{
# Define parameter ranges
par <- data.frame("snomelt_tmp.hru | change = absval" = c(-1, 1),
                 "snofall_tmp.hru | change = absval" = c(-1, 1))
par_center <- 2
par_oat <- sample_oat(par, par_center)
}
}
\seealso{
\code{\link{plot_oat}}
}
\keyword{sampling}
