% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sample_transect_i}
\alias{sample_transect_i}
\title{Sample Transect for a Specific Parameter}
\usage{
sample_transect_i(par_bnd, par_center, par_transect, i_c)
}
\arguments{
\item{par_bnd}{A list of parameter boundaries.}

\item{par_center}{A data frame or matrix of parameter centers.}

\item{par_transect}{A list of parameter transects.}

\item{i_c}{An integer indicating the index of the current parameter center.}
}
\value{
A data frame of samples along the transect for the specified parameter.
}
\description{
Generates samples along the transect for a specific parameter.
}
\keyword{internal}
