% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{sum_variables}
\alias{sum_variables}
\title{Sum values of simulated variables}
\usage{
sum_variables(...)
}
\arguments{
\item{...}{Variable tables which should be summed up.}
}
\value{
Returns a variable table with element wise summed values.
}
\description{
\code{sum_variables()} sums the values of provided simulated variable tables.
The \code{date} and the \code{unit} columns are excluded. All other elements are summed
element wise. This is useful e.g. to calculate total Nitrogen from simulated
N fractions.
}
\examples{
\dontrun{
no3_sim  <- sim$simulation$no3_day
nh3_sim  <- sim$simulation$nh3_day
no2_sim  <- sim$simulation$no2_day
orgn_sim <- sim$simulation$orgn_day

ntot_sim <- add_variables(no3_sim, nh3_sim, no2_sim, orgn_sim)
}

}
\keyword{helper}
