% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_helper.R
\name{translate_sequence}
\alias{translate_sequence}
\title{Translate the unit sequences to calibration.cal syntax}
\usage{
translate_sequence(val_seq)
}
\arguments{
\item{val_seq}{Numeric vector with the unit values}
}
\description{
Translate the unit sequences to calibration.cal syntax
}
\keyword{internal}
