% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{translate_to_boundaries}
\alias{translate_to_boundaries}
\title{Translate the normalized sampled values of a parameter to ranges of groups
of that parameter (e.g. when perco should have different ranges based on the
initial values of perco in all HRUs)}
\usage{
translate_to_boundaries(par_tbl, par_name, par_bound, par_group)
}
\arguments{
\item{par_tbl}{Table with the sampled parameter combinations}

\item{par_name}{Name of the parameter which should be translated}

\item{par_bound}{List of parameter boundaries to which the normalized values
should be transformed to.}

\item{par_group}{(named) Integer vector which defines the groups to which each
spatial unit belongs to.}
}
\value{
The table with the parameter samples, where the normalized parameter
is replaced by several columns with the updated parameter boundaries and
correct names assigned, so that the parameter boundaries are correctly applied.
}
\description{
Translate the normalized sampled values of a parameter to ranges of groups
of that parameter (e.g. when perco should have different ranges based on the
initial values of perco in all HRUs)
}
