% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{truncate}
\alias{truncate}
\title{Truncate a Vector with Ellipses}
\usage{
truncate(x, n, side = "left")
}
\arguments{
\item{x}{A vector to be truncated.}

\item{n}{An integer specifying the maximum number of elements to include before truncation.}

\item{side}{A string indicating the side from which to truncate. Options are \code{'left'} or \code{'both'}.
Default is \code{'left'}.}
}
\value{
A string with the truncated vector. If truncation occurs, ellipses (\code{"..."}) are added
to indicate omitted values.
}
\description{
This function truncates a vector and adds ellipses (\code{"..."}) to indicate omitted values.
Truncation can be done from the left or both sides.
}
\keyword{internal}
