% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{view_timeseries}
\alias{view_timeseries}
\title{View interactive time series of simulation results and observations}
\usage{
view_timeseries(
  sim,
  obs = NULL,
  run_ids = NULL,
  run_sel = NULL,
  plot_bands = TRUE,
  period = NULL,
  fn_summarize = "mean",
  x_label = "Date",
  y_label = "Discharge (m<sup>3</sup> s<sup>-1</sup>)"
)
}
\arguments{
\item{sim}{Data frame with one date column and one or many columns with
values of the simulated variable.}

\item{obs}{(optional) Data frame with one date and one value column.
Default \code{obs = NULL}, only simulations are plotted.}

\item{run_ids}{(optional) Integer vector of run IDs to plot.
Default \code{run_ids = NULL} requires input of run_sel.}

\item{run_sel}{(optional) Integer vector of run IDs to emphasize in the plot.
Default \code{run_sel = NULL} requires input of run_ids.}

\item{plot_bands}{(optional) Logical. If TRUE, includes lower and upper bands of results.
Default \code{plot_bands = TRUE}.}

\item{period}{(optional) Character defining the time step for aggregation. Options include "day", "week",
"month", "year", etc. Use "average monthly" for multi-annual monthly values.
Default \code{period = NULL} provides no aggregation. See \href{https://www.rdocumentation.org/packages/lubridate/versions/1.3.3/topics/floor_date}{lubridate::floor_date} for details.}

\item{fn_summarize}{Function to summarize time intervals. Default \code{fn_summarize ="mean"}, other examples are "median", "sum".
See \href{https://dplyr.tidyverse.org/reference/summarise.html}{dplyr::summarise} for details.}

\item{x_label}{(optional) Character for x axis label. Default \code{x_label = "Date"}.}

\item{y_label}{(optional) Character for y axis label. Default \code{y_label = "Discharge (m³/s)"}.}
}
\value{
Interactive plot of simulated and observed data.
}
\description{
This function plots interactive time series comparing simulated data with observed data.
}
\examples{
\dontrun{
view_timeseries(sim_flow, obs = obs, run_ids = run_sel_ids, run_sel = c(95),
plot_bands = TRUE, period = "average monthly")
}
}
\keyword{plot}
