% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_cal_file}
\alias{write_cal_file}
\title{Write Calibration File}
\usage{
write_cal_file(par, write_path = NULL, i_run = 1)
}
\arguments{
\item{par}{A data frame containing parameter information.}

\item{write_path}{(optional) The path where the calibration file will be written.
Default \code{write_path = NULL}, the current working directory will be used.}

\item{i_run}{(optional) An integer specifying the run number.
Default \code{i_run = 1}.}
}
\value{
None, but write out calibration.cal file
}
\description{
This function arranges and writes a calibration.cal file based on the provided
parameter table.
}
\examples{
\dontrun{
# Example usage:
write_cal_file(my_parameter_table, "path/to/write", 1)
}
}
\keyword{write}
