% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_tbl}
\alias{write_tbl}
\title{Write SWAT+ input file, which has a tabular structure.}
\usage{
write_tbl(tbl, file_path, fmt)
}
\arguments{
\item{tbl}{SWAT input table in tibble (data.frame) format.}

\item{file_path}{Write path of the SWAT+ input file.}

\item{fmt}{Character vector of format strings to define the print format of
each table column.}
}
\value{
Writes a text file table in the file path.
}
\description{
Write SWAT+ input file, which has a tabular structure.
}
\examples{
\dontrun{
## Read the hydrology.hyd input file
hydrology_hyd <- read_tbl(paste0(model_path, '/hydrology.hyd'))
hydrology_hyd$esco <- 1
## Updating hydrology.hyd file
hydr_hyd_fmt <- c('\%-16s', rep('\%12.5f', 14))
write_tbl(hydrology_hyd, paste0(model_path, '/hydrology.hyd'), hydr_hyd_fmt)
}
}
\seealso{
\code{\link{read_tbl}}
}
\keyword{write}
