% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{calc_gof}
\alias{calc_gof}
\title{Calculate a goodness-of-fit table}
\usage{
calc_gof(sim, obs, funs)
}
\arguments{
\item{sim}{Data frame with one date column and one or many columns with
values of the simulated variable.}

\item{obs}{Data frame with one date and one value column.}

\item{funs}{List of functions which are applied to calculate the
goodnes-of-fit values. The list can be a named list
(e.g. \code{list(nse_q = NSE, pb_q = pbias)}). In this case the column names of
returned table will be those names. If the list is unnamed then the function
names will be the column names.}
}
\value{
A table with the calculated goodness-of-fit values.
}
\description{
\code{calc_gof()} calculates goodness-of-fit inices for simulated time series
\code{sim} and observations \code{obs} applying the functions passed as a list
with \code{funs}.
}
\examples{
\dontrun{
date_flow <- seq(as.Date('2000-01-01'), as.Date('2000-12-31'), by = 'day')

flow_sim <- data.frame(date = date_flow,
                       run_1 = runif(length(date_flow)),
                       run_2 = runif(length(date_flow))
                       )
flow_obs <- data.frame(date = date_flow,
                       qobs = runif(length(date_flow))
                       )

obj_tbl <- calc_gof(sim = flow_sim, obs = flow_obs, funs = list(cor_q = cor))
}

}
\keyword{calculate}
