% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dotty_fig}
\alias{dotty_fig}
\title{Make a dotty plot}
\usage{
dotty_fig(sim_yield, yld_obs = NULL)
}
\arguments{
\item{sim_yield}{a dataframe with yield data}

\item{yld_obs}{(optional) The yield values for each crop with columns:
plant_name, yield_min, yield_max, yield_mean. Default \code{yld_obs = NULL}.}
}
\value{
ggplot object for dotty plot
}
\description{
Make a dotty plot
}
\examples{
\dontrun{
dotty_fig(sim_yield)
}
}
\keyword{internal}
\keyword{plot}
