% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dotty_yields}
\alias{plot_dotty_yields}
\title{Plot Dotty Yields}
\usage{
plot_dotty_yields(sim_result, yield_obs = NULL)
}
\arguments{
\item{sim_result}{The simulation results containing yield data.}

\item{yield_obs}{(optional) Data.frame with observed min, max, and mean yield
values for each crop. Default \code{yield = NULL}.}
}
\value{
A combined ggplot object showing dottty plots for 4 parameters.
}
\description{
This function creates dotty plots for simulated yields, comparing them with
observed yields (if available).
}
\examples{
\dontrun{
plot_phu_yld_bms(ylds_phu_dmat, dmat_chg, yield_mean, yield_min, yield_max)
}
}
\keyword{plot}
