% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_parameter_identifiability}
\alias{plot_parameter_identifiability}
\title{Plot Parameter Identifiability}
\usage{
plot_parameter_identifiability(parameters, objectives, run_fraction = NULL)
}
\arguments{
\item{parameters}{A vector of parameter names.}

\item{objectives}{A list of objective functions, each represented as a numeric vector.}

\item{run_fraction}{A numeric value representing the fraction of runs to consider for threshold calculation. Default is NULL.}
}
\value{
A ggplot object visualizing the identifiability of parameters.
}
\description{
This function creates a plot to visualize the identifiability of parameters
based on objective functions.
}
